# contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Internal Audience & Campaign Control API
  version: 0.8.28
  description: >
    Internal-only API for organisation records, offline consent reviews, secure audience imports,
    governed cohort definitions and controlled campaign lifecycle management.
servers:
  - url: /api/v1
security:
  - internalSession: []
paths:
...
  /internal/sender-nodes/{id}/heartbeat:
    post:
      operationId: heartbeatSenderNode
      summary: heartbeatSenderNode
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /internal/sender-sessions/{id}/heartbeat:
    post:
      operationId: heartbeatSenderSession
      summary: heartbeatSenderSession
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
...
  /sender-nodes:
    get:
      operationId: listSenderNodes
      summary: listSenderNodes
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Request completed
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
    post:
      operationId: registerSenderNode
      summary: registerSenderNode
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /sender-pools:
    get:
      operationId: listSenderPools
      summary: listSenderPools
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Request completed
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
    post:
      operationId: createSenderPool
      summary: createSenderPool
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /sender-pools/{id}:
    put:
      operationId: updateSenderPool
      summary: updateSenderPool
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /sender-pools/{id}/capacity:
    get:
      operationId: getSenderPoolCapacity
      summary: getSenderPoolCapacity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Request completed }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
  /sender-sessions:
    get:
      operationId: listSenderSessions
      summary: listSenderSessions
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Request completed
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
    post:
      operationId: registerSenderSession
      summary: Register a governed sender session with complete operational metadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SenderSessionRegistrationRequest' }
      responses:
        '201': { description: Sender session registered }
        '400': { description: Invalid request }
        '401': { description: Authentication required }
        '403': { description: Permission denied }
        '422': { description: Sender metadata, MSISDN, engine or capacity is invalid }
  /sender-sessions/{id}:
    delete:
      operationId: deleteGatewaySenderSession
      summary: Delete and retire an inactive governed gateway session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SenderLifecycleAction' }
      responses:
        '200': { description: Session deleted and retired }
        '403': { description: Recent MFA step-up required }
        '409': { description: Session owns work or changed concurrently }
  /sender-sessions/{id}/metadata:
    put:
      operationId: updateSenderSessionMetadata
      summary: Update governed sender ownership, registration, display and recovery-reference metadata
      description: Requires sender administration permission and recent MFA step-up. Recovery information is accepted only as a non-secret reference and is never returned.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SenderOperationalMetadataRequest' }
      responses:
        '200': { description: Sender metadata updated without exposing the recovery reference }
        '403': { description: Recent MFA step-up required }
        '404': { description: Sender session not found }
        '409': { description: Optimistic version conflict }
        '422': { description: Sender operational metadata is invalid }
  /sender-sessions/{id}/gateway/create:
    post:
      operationId: createGatewaySenderSession
      summary: Create the concrete gateway session for a governed sender
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SenderLifecycleAction' } } }
      responses: { '200': { description: Gateway session created } }
...
  /sender-sessions/{id}/pairing/qr:
    get:
      operationId: getGatewaySenderSessionQR
      summary: Obtain QR pairing material for a session in pairing state; technical administration and recent MFA required
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: QR pairing material }
        '403': { description: Technical-administrator permission or recent MFA step-up required }
  /sender-sessions/{id}/pairing/code:
    post:
      operationId: getGatewaySenderSessionPairingCode
      summary: Obtain a pairing code for a governed session; technical administration and recent MFA required
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [phoneNumber]
              properties: { phoneNumber: { type: string, pattern: '^\\+?[0-9]{8,15}$' } }
      responses:
        '200': { description: Pairing code material }
        '403': { description: Technical-administrator permission or recent MFA step-up required }
  /sender-sessions/{id}/start:
    post:
      operationId: startGatewaySenderSession
      summary: Start a paired or recoverable gateway session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SenderLifecycleAction' } } } }
      responses: { '200': { description: Session start accepted; readiness awaits heartbeat } }
  /sender-sessions/{id}/drain:
    post:
      operationId: drainGatewaySenderSession
      summary: Drain a ready or busy session from new allocation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SenderLifecycleAction' } } } }
      responses: { '200': { description: Session draining } }
...
components:
  securitySchemes:
    internalSession:
      type: apiKey
      in: cookie
      name: campaign_session
...
  schemas:
...
    SenderOperationalMetadataRequest:
      type: object
      additionalProperties: false
      required: [ownerReference, registrationCountryIso2, profileDisplayName, recoveryReference, expectedVersion, reason]
      properties:
        ownerReference: { type: string, minLength: 2, maxLength: 200 }
        registrationCountryIso2: { type: string, pattern: '^[A-Z]{2}$' }
        profileDisplayName: { type: string, minLength: 1, maxLength: 160 }
        recoveryReference: { type: string, minLength: 3, maxLength: 500, writeOnly: true, description: Non-secret reference to controlled recovery information; never returned. }
        expectedVersion: { type: integer, format: int64, minimum: 1 }
        reason: { type: string, minLength: 8, maxLength: 1000 }
...
    SenderLifecycleAction:
      type: object
      additionalProperties: false
      required: [expectedVersion, reason]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 1 }
        reason: { type: string, minLength: 8 }
...